--- src/main/java/domain/unit/NoneUnitRule.java
@@ -1,17 +1,0 @@
-package domain.unit;
-
-import domain.position.Position;
-import domain.position.Route;
-import java.util.List;
-
-public class NoneUnitRule implements UnitRule {
-    @Override
-    public List<Route> calculateAllRoute(Position start) {
-        return List.of();
-    }
-
-    @Override
-    public UnitType getType() {
-        return UnitType.NONE;
-    }
-}
