--- src/main/java/domain/unit/Team.java
@@ -1,7 +1,0 @@
-package domain.unit;
-
-public enum Team {
-    HAN,
-    CHO,
-    ;
-}
