--- src/main/java/domain/unit/UnitRule.java
@@ -1,11 +1,0 @@
-package domain.unit;
-
-import domain.position.Position;
-import domain.position.Route;
-import java.util.List;
-
-public interface UnitRule {
-    List<Route> calculateAllRoute(Position start);
-
-    UnitType getType();
-}
