--- src/main/java/domain/unit/UnitType.java
@@ -1,13 +1,0 @@
-package domain.unit;
-
-public enum UnitType {
-    KING,
-    SCHOLAR,
-    CAR,
-    HORSE,
-    ELEPHANT,
-    BOMB,
-    JOL,
-    NONE,
-    ;
-}
