+++ src/main/java/janggi/game/GameState.java
@@ -1,0 +1,5 @@
+package janggi.game;
+
+public enum GameState {
+    PLAY, QUIT
+}
