+++ src/main/java/janggi/position/Point.java
@@ -1,0 +1,23 @@
+package janggi.position;
+
+public class Point {
+    private final int x;
+    private final int y;
+
+    private Point(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public static Point of(int x, int y) {
+        return new Point(x, y);
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public int getY() {
+        return y;
+    }
+}
