+++ src/main/java/janggi/unit/Team.java
@@ -1,0 +1,7 @@
+package janggi.unit;
+
+public enum Team {
+    HAN,
+    CHO,
+    ;
+}
