+++ src/main/java/janggi/unit/Unit.java
@@ -1,0 +1,61 @@
+package janggi.unit;
+
+import janggi.position.Position;
+import janggi.position.Route;
+import java.util.List;
+import java.util.Objects;
+
+public class Unit {
+    private final Team team;
+    private final UnitRule unitRule;
+    private Position position;
+
+    public Unit(Position position, Team team, UnitRule unitRule) {
+        this.position = position;
+        this.team = team;
+        this.unitRule = unitRule;
+    }
+
+    public static Unit of(Position position, Team team, UnitRule unitRule) {
+        return new Unit(position, team, unitRule);
+    }
+
+    public List<Route> calculateRoutes() {
+        return unitRule.calculateAllRoute(this.position);
+    }
+
+    public boolean isSamePoint(Position position) {
+        return this.position.equals(position);
+    }
+
+    public UnitType getType() {
+        return this.unitRule.getType();
+    }
+
+    public Team getTeam() {
+        return team;
+    }
+
+    public Position getPosition() {
+        return this.position;
+    }
+
+    public void move(Position position) {
+        this.position = position;
+    }
+
+    @Override
+    public boolean equals(Object object) {
+        if (object == null || getClass() != object.getClass()) {
+            return false;
+        }
+        Unit unit = (Unit) object;
+        return team == unit.team && Objects.equals(unitRule, unit.unitRule) && Objects.equals(position,
+                unit.position);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(team, unitRule, position);
+    }
+}
