+++ src/main/java/janggi/unit/UnitRule.java
@@ -1,0 +1,11 @@
+package janggi.unit;
+
+import janggi.position.Position;
+import janggi.position.Route;
+import java.util.List;
+
+public interface UnitRule {
+    List<Route> calculateAllRoute(Position start);
+
+    UnitType getType();
+}
