+++ src/main/java/janggi/unit/UnitType.java
@@ -1,0 +1,13 @@
+package janggi.unit;
+
+public enum UnitType {
+    KING,
+    SCHOLAR,
+    CAR,
+    HORSE,
+    ELEPHANT,
+    BOMB,
+    JOL,
+    NONE,
+    ;
+}
