+++ src/main/java/janggi/view/ConsoleColors.java
@@ -1,0 +1,10 @@
+package janggi.view;
+
+public final class ConsoleColors {
+    public static final String HAN_COLOR = "\u001B[31m";
+    public static final String CHO_COLOR = "\u001B[34m";
+    public static final String RESET_COLOR = "\u001B[0m";
+
+    private ConsoleColors() {
+    }
+}
